+++ src/runtime/virtcontainers/mount_darwin.go
@@ -1,0 +1,12 @@
+// Copyright (c) 2023 Apple Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+
+package virtcontainers
+
+import "context"
+
+func nydusContainerCleanup(ctx context.Context, sharedDir string, c *Container) error {
+	return nil
+}
